requirements.txt
Flask
Flask-Session
Flask-SQLAlchemy
Werkzeug
psycopg2-binary
python-dotenv
requests
gunicorn